requirements.txt
absl-py==0.9.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
future==0.18.2
gitdb==4.0.4
GitPython==3.1.1
google-auth==1.11.2
google-auth-oauthlib==0.4.1
grpcio==1.27.2
gym==0.16.0
-e git+https://github.com/maximecb/gym-minigrid.git@a5f4ef3684875eba613087dcef21be57c186bdc8#egg=gym_minigrid
-e git+https://github.com/maximecb/gym-miniworld.git@7531d6ca5a4de2361f6d00d1b24bfa135075736b#egg=gym_miniworld
idna==2.9
joblib==0.14.1
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.0.3
numpy==1.18.1
oauthlib==3.1.0
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
smmap==3.0.2
tensorboard==2.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.42.1
urllib3==1.25.8
Werkzeug==1.0.0
